First working version

+++ PITL.Extract.Job.Abstractions/Position.cs
@@ -5,6 +5,6 @@ namespace PITL.Extract.Job.Abstractions
     public record Position
     (
         DateTime LocalTime,
-        int Volume
+        double Volume
     );
 }
